move module home screen
