+++ ui/src/main/java/com/jin/ui/home/content/FoodSearch.kt
@@ -1,4 +1,4 @@
-package com.jin.honey.feature.home.ui.content
+package com.jin.ui.home.content
 
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.animation.fadeIn
@@ -30,7 +30,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import com.jin.honey.R
+import com.jin.ui.R
 import com.jin.domain.food.model.MenuPreview
 import com.jin.ui.theme.FoodSearchBoxBorderColor
 import com.jin.ui.theme.FoodSearchBoxTextColor
